Renamed plant api file + updated references + added ApiBase superclass
